% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{build_alternate_urls}
\alias{build_alternate_urls}
\title{Build alternate URLs to try if primary URL fails}
\usage{
build_alternate_urls(end_year, school_year, level, data_type, count_day)
}
\arguments{
\item{end_year}{School year end}

\item{school_year}{School year string}

\item{level}{"school" or "district"}

\item{data_type}{"grade" or "demo"}

\item{count_day}{"45", "135", or "180"}
}
\value{
Vector of alternate URLs
}
\description{
Build alternate URLs to try if primary URL fails
}
\keyword{internal}
