% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{build_headcount_url}
\alias{build_headcount_url}
\title{Build URL for headcount file}
\usage{
build_headcount_url(end_year, school_year, level, data_type, count_day)
}
\arguments{
\item{end_year}{School year end}

\item{school_year}{School year string}

\item{level}{"school" or "district"}

\item{data_type}{"grade" or "demo"}

\item{count_day}{"45", "135", or "180"}
}
\value{
URL string
}
\description{
URL patterns have changed over the years. This function attempts to build
the correct URL for a given year.
}
\keyword{internal}
