% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_headcount_file}
\alias{download_headcount_file}
\title{Download a single headcount file from SCDE}
\usage{
download_headcount_file(end_year, school_year, level, data_type, count_day)
}
\arguments{
\item{end_year}{School year end}

\item{school_year}{School year string (e.g., "2024-25")}

\item{level}{"school" or "district"}

\item{data_type}{"grade" or "demo"}

\item{count_day}{"45", "135", or "180"}
}
\value{
Data frame
}
\description{
Download a single headcount file from SCDE
}
\keyword{internal}
