% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_demo_column_names}
\alias{get_demo_column_names}
\title{Get column names for demographic data files}
\usage{
get_demo_column_names(level)
}
\arguments{
\item{level}{"school" or "district"}
}
\value{
Vector of column names
}
\description{
Get column names for demographic data files
}
\keyword{internal}
