% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{get_sc_district_codes}
\alias{get_sc_district_codes}
\title{Look up district ID from master list}
\usage{
get_sc_district_codes()
}
\value{
Named vector of district IDs
}
\description{
SC Department of Education assigns numeric codes to districts.
This function maps district names to official codes when available.
}
\keyword{internal}
