% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{process_directory}
\alias{process_directory}
\title{Process raw SC directory data to standard schema}
\usage{
process_directory(raw_data, end_year)
}
\arguments{
\item{raw_data}{Raw data frame from get_raw_directory()}

\item{end_year}{School year end}
}
\value{
Processed tibble with standard schema
}
\description{
Takes raw directory data from the SC Report Cards 1.MainPage sheet and
standardizes column names and types.
}
\keyword{internal}
