% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_district_enr}
\alias{process_district_enr}
\title{Process district-level enrollment data}
\usage{
process_district_enr(grade_df, demo_df, end_year)
}
\arguments{
\item{grade_df}{District grade data frame}

\item{demo_df}{District demographic data frame}

\item{end_year}{School year end}
}
\value{
Processed district data frame
}
\description{
Merges grade and demographic data for districts.
}
\keyword{internal}
