% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_enr}
\alias{process_enr}
\title{Process raw SCDE enrollment data}
\usage{
process_enr(raw_data, end_year)
}
\arguments{
\item{raw_data}{List containing school_grade, school_demo, district_grade,
district_demo data frames from get_raw_enr}

\item{end_year}{School year end}
}
\value{
Processed data frame with standardized columns
}
\description{
Transforms raw Active Student Headcount data into a standardized schema
combining school and district data.
}
\keyword{internal}
