% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_school_enr}
\alias{process_school_enr}
\title{Process school-level enrollment data}
\usage{
process_school_enr(grade_df, demo_df, end_year)
}
\arguments{
\item{grade_df}{School grade data frame}

\item{demo_df}{School demographic data frame}

\item{end_year}{School year end}
}
\value{
Processed school data frame
}
\description{
Merges grade and demographic data for schools.
}
\keyword{internal}
